+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2026-02-22] - Option for Multiple Search Queries
+
+### Added
+
+- SearchEngine now supports multiple search queries - it narrows down results with each additional query.
+
+---
+
 ## [2026-02-17] - description made optional, library list update
 
 ### Changed